Create phonebook.sh
This was one of my first "big" unix projects. A simple phonebook directory which stores entries from the user and provides (s)he to view/search/delete previously stored entries from the file. I have refactored the code a bit, but its almost exactly the same as the original (so as to respect the work of my younger self).